Gradually increase Gen size in `sample`
At the moment, the `Size` parameter of `sample` is used not only for the
number of values that you want to get back, but also for the size
parameter for the `GenState`. This is not ideal, as often you want to
change one without affecting the other.

This commit changes this behaviour so that the parameter is now written
as `Int` instead, and affects only the number of values which are
produced. The Size parameter now starts at 0 and is incremented in
steps of 2, just like in the Haskell version.